requirements.txt
colorama>=0.4.6
selenium>=4.10.0
pyodbc>=4.0.40
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.1.0
geopandas>=0.19.0
rasterio>=1.3.9
folium>=0.16.0
matplotlib>=3.8.0
